==> Scripts/03_EqSim_summarize PELAC.Rmd <==
---
output: 
  word_document:
    reference_docx: ..\PFA_report_template_v1.4.dotx
---

```{r setup, include=FALSE}

################################################################################
# EqSim Summarize.Rmd
#
# 06/07/2020 tested on 1000 iters of SAM assessment
# 05/08/2020 converted to Rmd file
#
################################################################################

require("knitr")
knitr::opts_chunk$set(echo = FALSE, message = FALSE, warning = FALSE, crop = TRUE, comment = "")
knitr::opts_chunk$set(fig.width=10) 

options(dplyr.summarise.inform = FALSE)


# rm(list=ls())
invisible(gc())

library(FLCore)
library(tidyverse)


Base.dir <- dirname(getwd())
MSE.dir   <- file.path(Base.dir, "Scripts")
Data.dir  <- file.path(Base.dir,"data")              
RData.dir <- Data.dir
Res.dir   <- file.path(Base.dir, "Results")
stats.dir <- Res.dir
wkrebuild.dir <- "C:/DATA/GIT/wk_WKREBUILD/EqSimWHM/Results/Stats"

Source.dir <- file.path(Base.dir,"R")              #R functions
source(file.path(Source.dir,"utilities.R"))

# Rebuilding threshold
rebuiltThreshold <- 0.5

file.list <- list.files(path=stats.dir, pattern="_Stats", full.names=TRUE)

# Add results from WKWHMRP analysis
file.list <- c(file.list,
               file.path(wkrebuild.dir, "WHOM_SAM_2020_OM2.5_1000_MP5.23_23_eqSim_Stats.RData"),
               file.path(wkrebuild.dir, "WHOM_SS3_2020_OM2.3_1000_MP5.23_23_eqSim_Stats.RData")
               )

# manual fixing of units
my_units <- data.frame(
  perfstat = c("stock","fbar","catch","rec",     "tac", "pblim", "precbpa", "iav"),
  unit2     = c(   "kt","F(1-10)",   "kt","billions", "kt", "proportion","proportion", "proportion change"),
  stringsAsFactors = FALSE)

# print(file.list)
i <- 1

df <- rp <- data.frame(stringsAsFactors = FALSE)

for (i in 1:length(file.list)) {
  print(file.list[i])
  df <- bind_rows(df, loadRData(file.list[i])[["df"]])
  rp <- bind_rows(rp, 
                  loadRData(file.list[i])[["OM"]][["refPts"]] %>% 
                    as.data.frame() %>% 
                    mutate(assess     = stringr::word(basename(file.list[i]), 2, sep="_")) %>% 
                    mutate(assessyear = stringr::word(basename(file.list[i]), 3, sep="_")) %>% 
                    lowcase() %>% 
                    mutate(flow = fmsy/5) %>% 
                    tidyr::pivot_longer(names_to = "refpoint", values_to = "data", 
                                        c("fpa","flim","fmsy","flow", "msybtrigger", "blim", "bpa","bloss"))  %>% 
                    mutate(perfstat = ifelse(grepl("^f", refpoint), "fbar","stock")) %>% 
                    mutate(data = ifelse(perfstat=="stock", data/1000, data)) %>% 
                    tidyr::pivot_wider(names_from = refpoint, values_from = data)
                  )
                    
}

# loadRData(file.list[3])[["df"]] %>% View()



# COULD BE DONE IN SIMULATION CODE
df <- 
  df %>%  
  
  filter(niters==1000) %>% 

  mutate(method = "EQSIM") %>% 
  
  # mutate(period = ifelse(is.na(period) & year <= (an(assessyear)), "HI",period)) %>% 
  mutate(assess = ifelse(assess =="SS", "SS3", assess) ) %>% 
  mutate(period = ifelse(an(year) < an(assessyear), "HI",period) ) %>% 
  mutate(ftgt = ac(ftgt)) %>% 
  
  left_join(my_units, by="perfstat") %>% 
  
  mutate(
    mean   = ifelse(perfstat == "stock", mean/1000, mean),
    median = ifelse(perfstat == "stock", median/1000, median),
    lower  = ifelse(perfstat == "stock", lower/1000, lower),
    upper  = ifelse(perfstat == "stock", upper/1000, upper),
    data   = ifelse(perfstat == "stock", data/1000, data),
  ) %>% 
  mutate(
    blim      = ifelse(perfstat == "stock", blim/1000, NA),
    bpa       = ifelse(perfstat == "stock", bpa/1000, NA),
    msybtrig  = ifelse(perfstat == "stock", bpa/1000, NA)
  ) %>%  
  
  mutate(
    mean   = ifelse(perfstat == "rec", mean/1000000, mean),
    median = ifelse(perfstat == "rec", median/1000000, median),
    lower  = ifelse(perfstat == "rec", lower/1000000, lower),
    upper  = ifelse(perfstat == "rec", upper/1000000, upper),
    data   = ifelse(perfstat == "rec", data/1000000, data)
  ) %>% 

  mutate(
    mean   = ifelse(perfstat == "catch", mean/1000, mean),
    median = ifelse(perfstat == "catch", median/1000, median),
    lower  = ifelse(perfstat == "catch", lower/1000, lower),
    upper  = ifelse(perfstat == "catch", upper/1000, upper),
    data   = ifelse(perfstat == "catch", data/1000, data)
  ) 


# save(df, file="df.RData")

# ========================================================================================================

# plotvar function

plotvar <- function(mystock      =  "WHOM",
                    myassess     = "SAM",
                    myassessyear = c("2019","2020"),
                    mymethod  = "EQSIM",
                    myom         = c("OM2.4","OM2.5"),
                    myniters     = "1000",
                    mymp         = c("MP5.23"),
                    mynyrs       = "23",
                    myftgt       = ac(c(0.0, 0.025, 0.05, 0.075, 0.10, 0.125, 0.15)),
                    myperfstat   = "stock",
                    mycolour     = "blue",
                    myyintercept = "blim",
                    myvalue      = "median",
                    myfacets     = c("assessyear","ftgt"),
                    myfirstyear  = as.numeric(NA),
                    mylastyear   = as.numeric(NA),
                    scale_y_free = FALSE) {
  
  
  # history
  h <-
    df %>% 
    filter(period     %in% c("HI")) %>% 
    filter(perfstat   %in% myperfstat) %>% 
    filter(assess     %in% myassess) %>% 
    filter(assessyear %in% myassessyear) %>% 
    filter(method     %in% mymethod) %>% 
    filter(om         %in% myom) %>% 
    filter(mp         %in% mymp) %>% 
    # filter(nyrs       %in% mynyrs) %>% 
    filter(ftgt       %in% ac(myftgt)) %>% 
    
    {if(!is.na(myfirstyear)) filter(., year >= myfirstyear) else (.)} %>%    
    {if(!is.na(mylastyear))  filter(., year <= mylastyear) else (.)} %>%   
    
    mutate(age = as.numeric(age)) %>% 
    mutate(assess = factor(assess, levels=myassess))

  # future
  d <- 
    df %>%
    filter(period     %in% c("CU","ST","MT",'LT')) %>% 
    filter(perfstat   %in% myperfstat) %>% 
    filter(stock      %in% mystock) %>% 
    filter(assess     %in% myassess) %>% 
    filter(assessyear %in% myassessyear) %>% 
    filter(method  %in% mymethod) %>% 
    filter(om         %in% myom) %>% 
    filter(mp         %in% mymp) %>%
    # filter(nyrs       %in% mynyrs) %>% 
    filter(ftgt       %in% ac(myftgt)) %>% 
    
    {if(!is.na(myfirstyear)) filter(., year >= myfirstyear) else (.)} %>%    
    {if(!is.na(mylastyear))  filter(., year <= mylastyear) else (.)} %>%   
    
    mutate(age  = as.numeric(age)) %>% 
    mutate(assess = factor(assess, levels=myassess))

    
  # periods
  p <-
    bind_rows(h, d) %>% 
    distinct(assessyear, period, year) %>% 
    
    {if(!is.na(myfirstyear)) filter(., year >= myfirstyear) else (.)} %>%    
    {if(!is.na(mylastyear))  filter(., year <= mylastyear) else (.)} %>%   
    
    group_by(assessyear, period) %>% 
    summarise(
      minyear = min(year),
      maxyear = max(year)
    )
  
  # intercept
  if(!all(is.na(myyintercept))) {
    i <-
      h %>% 
      distinct(perfstat, y=get(myyintercept)) %>% 
      drop_na() 
  } else {
    i <- an(NA)
  }

  # title
  t <- paste(
    toupper(mystock),
    paste(unique(myassess), collapse="-"),
    paste(unique(myassessyear), collapse="-"),
    # paste(unique(mymethod), collapse="-"),
    paste(unique(myom), collapse="-"),
    paste(unique(mymp), collapse="-"),
    toupper(myperfstat),
    sep="_"
  )
  
  # h %>% filter(!is.na(iter), !is.na(data)) %>% View()
  
  myfig <- 

    ggplot(d, aes(x=year, y=get(myvalue), group=mp)) +
    theme_publication() +
    theme(axis.text.x = element_text(angle = 90, vjust=0.5)) +
    theme( panel.grid.major.x = element_blank()) +
    theme(legend.position = "none") +
    theme(panel.spacing = unit(0.2, "lines")) +
    
    # historical ribbon
    geom_ribbon(data=filter(h, !is.na(get(myvalue))),
                aes(ymin=lower, ymax=upper, group=iter), alpha=0.4, fill="gray") +
    
    # historical worms
    geom_line(data=filter(h, metric=="worm"),
              aes(x=year, y=data, group=iter), colour="darkgray", size=0.5) +
    
    # historical median
    geom_line(data=filter(h, !is.na(get(myvalue))),
              aes(x=year,y=get(myvalue),group=iter), colour="black", size=1) +
    
    # future ribbon
    geom_ribbon(data=filter(d, !is.na(get(myvalue))),
                aes(ymin=lower, ymax=upper, group=iter, fill=perfstat), alpha=0.4) +
    scale_fill_manual(values=mycolour) + 
    
    # future worms
    geom_line(data=filter(d, metric=="worm"),
              aes(x=year, y=data, group=iter, colour=perfstat), size=0.5) +
    scale_colour_manual(values=mycolour) + 
    
    # future median
    geom_line(data=filter(d, !is.na(get(myvalue))),
              aes(x=year,y=get(myvalue),group=iter, colour=perfstat), size=1) +
    
    # hline
    {if(!all(is.na(myyintercept))) geom_hline(data=i,
                                              aes(yintercept=y), linetype="dashed")}  +
    
    # vertical lines (periods)
    geom_vline(data=filter(p, period=="CU"),
               aes(xintercept=minyear), linetype="dotted") +
    geom_vline(data=filter(p, period=="ST"),
               aes(xintercept=minyear), linetype="dotted") +
    geom_vline(data=filter(p, period=="MT"),
               aes(xintercept=minyear), linetype="dotted") +
    geom_vline(data=filter(p, period=="LT"),
               aes(xintercept=minyear), linetype="dotted") +
    geom_vline(data=filter(p, period=="LT"),
               aes(xintercept=maxyear), linetype="dotted") +
    
    expand_limits(y=0) +
    labs(x="", y=myperfstat, title=t) +

    {if(scale_y_free==TRUE) {
      facet_grid(get(myfacets[1]) ~ get(myfacets[2]), scales="free_y")
    } else {
      facet_grid(get(myfacets[1]) ~ get(myfacets[2]))
    }}
    
  print(myfig)
  
  ggsave(file = file.path(Res.dir,paste0(t, ".png")),
         device="png", width = 30, height = 20, units = "cm")

}

# ========================================================================================================

# plotrecovery

# mystock      = "WHOM"
# myassess     = "SAM"
# myassessyear = c("2022")
# mymethod     = c("EQSIM")
# myom         = c("OM2.5")
# mymp         = c("MP5.23")
# myftgt       = c(0.001, 0.05, 0.075, 0.10, 0.15)
# myfacets     = c("mp","ftgt")
# myfirstyear  = 2005
# mylastyear   = max(df$year)
# myintercept  = 0.5

plotrecovery <- function(
                    mystock      =  "WHOM",
                    myassess     = "SAM",
                    myassessyear = c("2019","2020"),
                    mymethod     = "EQSIM",
                    myom         = c("OM2.4","OM2.5"),
                    mymp         = c("MP5.23"),
                    myftgt       = c(0.0, 0.025, 0.05, 0.075, 0.10, 0.125, 0.15),
                    myintercept  = 0.5,
                    myfacets     = c("assessyear","ftgt"),
                    myfirstyear  = as.numeric(NA),
                    mylastyear   = as.numeric(NA)) {
  
  # plot recovery to blim and bpa
  d <-
    df %>%
    filter(perfstat %in% c("precblim","precbpa")) %>%
    filter(stock      %in% mystock) %>% 
    filter(assess     %in% myassess) %>% 
    filter(assessyear %in% myassessyear) %>% 
    filter(method  %in% mymethod) %>% 
    filter(om         %in% myom) %>% 
    filter(mp         %in% mymp) %>% 
    filter(ftgt       %in% myftgt) %>% 
    
    {if(!is.na(myfirstyear)) filter(., year >= myfirstyear) else (.)} %>%    
    {if(!is.na(mylastyear))  filter(., year <= mylastyear) else (.)} %>% 
    
    mutate(assess = factor(assess, levels=myassess))

  
  # lines
  p <-
    df %>%
    filter(perfstat %in% c("firstyearrebuildtoblim","firstyearrebuildtobpa")) %>%
    filter(stock      %in% mystock) %>% 
    filter(assess     %in% myassess) %>% 
    filter(assessyear %in% myassessyear) %>% 
    filter(method  %in% mymethod) %>% 
    filter(om         %in% myom) %>% 
    filter(mp         %in% mymp) %>% 
    filter(ftgt       %in% myftgt) %>% 
    
    mutate(perfstat = case_when(
      perfstat == "firstyearrebuildtoblim" ~ "precblim",
      perfstat == "firstyearrebuildtobpa"  ~ "precbpa",
      TRUE                                 ~ perfstat
    )) %>% 
    
    mutate(assess = factor(assess, levels=myassess))


  # title
  t <- paste(
    toupper(mystock),
    paste(unique(myassess), collapse="-"),
    paste(unique(myassessyear), collapse="-"),
    # paste(unique(mymethod), collapse="-"),
    paste(unique(myom), collapse="-"),
    paste(unique(mymp), collapse="-"),
    "recovery",
    sep="_"
  )
  
  myfig <-
    d %>% 
    ggplot(aes(x=year, y=mean, group=perfstat)) +
    theme_publication() +
    theme(axis.text.x = element_text(angle = 90, vjust=0.5)) +
    theme( panel.grid.major.x = element_blank()) +
    
    geom_vline(data=p, aes(xintercept=mean, colour=perfstat), linetype="dashed") +

    geom_hline(yintercept=myintercept, colour="gray", linetype="dashed") +
    
    geom_text(data=filter(p, perfstat %in% c("precblim")),
              aes(x=mean, label=ac(mean), colour=perfstat),
              y=1.0, vjust=1, hjust=0, nudge_x = 1, size=3) +
    
    geom_text(data=filter(p, perfstat %in% c("precbpa")),
              aes(x=mean, label=ac(mean), colour=perfstat),
              y=0.0, vjust=0, hjust=0, nudge_x = 1, size=3) +
    
    geom_line(aes(colour=perfstat)) +
    
    expand_limits(y=0) +
    labs(x="", y="", title=t) +
    facet_grid(get(myfacets[1]) ~ get(myfacets[2])) 
  
  print(myfig)
  
  ggsave(file = file.path(Res.dir,paste0(t, ".png")),
         device="png", width = 30, height = 20, units = "cm")
  
  
} # end of plotrecovery

# ========================================================================================================

# plot prob being below blim

# mystock      = "WHOM"
# myassess     = c("SS3", "SAM")
# myassessyear = c("2020", "2022")
# mymethod     = c("EQSIM")
# myom         = c("OM2.3", "OM2.5")
# mymp         = c("MP5.23")
# myftgt       = c(0.074)
# myfacets     =  c("assess", "assessyear")
# myfirstyear  = 2005
# mylastyear   = NA
# mycolour = "red"

plotpblim <- function(
                    mystock      =  "WHOM",
                    myassess     = "SAM",
                    myassessyear = c("2019","2020"),
                    mymethod     = "EQSIM",
                    myom         = c("OM2.4","OM2.5"),
                    mymp         = c("MP5.23"),
                    myftgt       = c(0.0, 0.025, 0.05, 0.075, 0.10, 0.125, 0.15),
                    myfacets     = c("assessyear","ftgt"),
                    myfirstyear  = as.numeric(NA),
                    mylastyear   = as.numeric(NA),
                    mycolour     = "red") {
  
  # plot historical probability below blim and bpa
  h <-
    df %>% 
    # filter(period     %in% c("HI")) %>% 
    filter(perfstat   %in% c("pblim")) %>%
    filter(stock      %in% mystock) %>% 
    filter(assess     %in% myassess) %>% 
    filter(assessyear %in% myassessyear) %>% 
    filter(method     %in% mymethod) %>% 
    filter(om         %in% myom) %>% 
    filter(mp         %in% mymp) %>% 
    filter(ftgt       %in% myftgt) %>% 
    
    {if(!is.na(myfirstyear)) filter(., year >= myfirstyear) else (.)} %>%    
    {if(!is.na(mylastyear))  filter(., year <= mylastyear) else (.)}%>% 
    
    mutate(assess = factor(assess, levels=myassess))
    
  # plot probability below blim and bpa
  d <-
    df %>%
    filter(period     %in% c("CU","ST","MT",'LT')) %>% 
    filter(perfstat   %in% c("pblim")) %>%
    filter(stock      %in% mystock) %>% 
    filter(assess     %in% myassess) %>% 
    filter(assessyear %in% myassessyear) %>% 
    filter(method     %in% mymethod) %>% 
    filter(om         %in% myom) %>% 
    filter(mp         %in% mymp) %>% 
    filter(ftgt       %in% myftgt) %>% 
    
    {if(!is.na(myfirstyear)) filter(., year >= myfirstyear) else (.)} %>%    
    {if(!is.na(mylastyear))  filter(., year <= mylastyear) else (.)}%>% 
    
    mutate(assess = factor(assess, levels=myassess))

  
  # periods
  p <-
    d %>% 
    distinct(assessyear, period, year) %>% 
    
    {if(!is.na(myfirstyear)) filter(., year >= myfirstyear) else (.)} %>%    
    {if(!is.na(mylastyear))  filter(., year <= mylastyear) else (.)} %>%   
    
    group_by(assessyear, period) %>% 
    summarise(
      minyear = min(year),
      maxyear = max(year)
    )
  
  # firstyear pblim
  p1 <-
    d %>% 
    filter(mean       < 0.05) %>% 
    group_by(perfstat, stock, assess, assessyear, method, om, mp, ftgt) %>% 
    summarise(mean = min(year, na.rm=TRUE))

  # title
  t <- paste(
    toupper(mystock),
    paste(unique(myassess), collapse="-"),
    paste(unique(myassessyear), collapse="-"),
    # paste(unique(mymethod), collapse="-"),
    paste(unique(myom), collapse="-"),
    paste(unique(mymp), collapse="-"),
    "pBlim",
    sep="_"
  )
  
  myfig <-
    d %>% 
    ggplot(aes(x=year, y=mean, group=perfstat)) +
    theme_publication() +
    theme(axis.text.x = element_text(angle = 90, vjust=0.5)) +
    theme( panel.grid.major.x = element_blank()) +
    
    geom_hline(yintercept=0.05, colour="gray", linetype="dashed") +
    
    geom_vline(data=p1, aes(xintercept=mean), colour=mycolour, linetype="dashed") +
    geom_text(data= p1,
              aes(x=mean, label=ac(mean)),
              y=0.05, vjust=0, hjust=0, nudge_x = 0.2, colour=mycolour) +
    
    geom_line(data=h, colour="darkgray") +
    geom_line(aes(colour=perfstat), size=1) +
    
    # vertical lines (periods)
    geom_vline(data=filter(p, period=="CU"),
               aes(xintercept=minyear), linetype="dotted") +
    geom_vline(data=filter(p, period=="ST"),
               aes(xintercept=minyear), linetype="dotted") +
    geom_vline(data=filter(p, period=="MT"),
               aes(xintercept=minyear), linetype="dotted") +
    geom_vline(data=filter(p, period=="LT"),
               aes(xintercept=minyear), linetype="dotted") +
    geom_vline(data=filter(p, period=="LT"),
               aes(xintercept=maxyear), linetype="dotted") +

    expand_limits(y=0) +
    labs(x="", y="", title=t) +
    facet_grid(get(myfacets[[1]]) ~ get(myfacets[[2]])) 
  
  print(myfig)
  
  ggsave(file = file.path(Res.dir,paste0(t, ".png")),
         device="png", width = 30, height = 20, units = "cm")
  
  
} # end of plotpblim



# ========================================================================================================

# plotscenario

# mystock      =  "WHOM"
# myassess     = "SS3"
# myassessyear = c("2020")
# mymethod     = "EQSIM"
# myom         = c("OM2.3")
# mymp         = c("MP5.23")
# myftgt       = c(0.075)
# myfacets     = c("perfstat")
# myfirstyear  = 2000
# mylastyear   = as.numeric(NA)
# myperfstat   = c("stock","fbar", "pblim", "precbpa")
# mycolours    = c("red","blue","green", "blue")

# mystock      =  "WHOM"
# myassess     = "SS3"
# myassessyear = c("2020")
# mymethod     = "EQSIM"
# myom         = c("OM2.3")
# mymp         = c("MP5.23")
# myftgt       = c("0.074")
# myfacets     = c("perfstat")
# myfirstyear  = 2000
# mylastyear   = as.numeric(NA)
# myperfstat   = c("fbar")
# mycolours    = c("blue")
# mylastyear   = as.numeric(NA)
# showtitle = TRUE

plotscenario <- function(
                    mystock      =  "WHOM",
                    myassess     = "SAM",
                    myassessyear = c("2022"),
                    mymethod     = "EQSIM",
                    myom         = c("OM2.5"),
                    mymp         = c("MP5.25"),
                    myftgt       = c(0.075),
                    myfacets     = c("perfstat"),
                    myfirstyear  = as.numeric(NA),
                    mylastyear   = as.numeric(NA),
                    myperfstat   = c("stock","pblim"),
                    mycolours    = rep(c("red","blue","green", "blue"),5),
                    showtitle    = FALSE, 
                    ncolumns     = 1 ) {
  
  # history
  h <-
    df %>% 
    # filter(period     %in% c("HI")) %>% 
    filter(perfstat   %in% myperfstat) %>%
    filter(stock      %in% mystock) %>% 
    filter(assess     %in% myassess) %>% 
    filter(assessyear %in% myassessyear) %>% 
    filter(method     %in% mymethod) %>% 
    filter(om         %in% myom) %>% 
    filter(mp         %in% mymp) %>%
    filter(ftgt       %in% myftgt) %>% 
    
    {if(!is.na(myfirstyear)) filter(., year >= myfirstyear) else (.)} %>%    
    {if(!is.na(mylastyear))  filter(., year <= mylastyear) else (.)} %>% 
    
    mutate(age = as.numeric(age)) 

  # future
  d <- 
    df %>%
    filter(period     %in% c("CU","ST","MT",'LT')) %>% 
    filter(perfstat   %in% myperfstat) %>%
    filter(stock      %in% mystock) %>% 
    filter(assess     %in% myassess) %>% 
    filter(assessyear %in% myassessyear) %>% 
    filter(method  %in% mymethod) %>% 
    filter(om         %in% myom) %>% 
    filter(mp         %in% mymp) %>% 
    filter(ftgt       %in% myftgt) %>% 
    
    {if(!is.na(myfirstyear)) filter(., year >= myfirstyear) else (.)} %>%    
    {if(!is.na(mylastyear))  filter(., year <= mylastyear) else (.)} %>% 
    
    mutate(age  = as.numeric(age)) 
    
  # periods
  p <-
    bind_rows(h, d) %>% 
    distinct(assessyear, period, year) %>% 
    
    {if(!is.na(myfirstyear)) filter(., year >= myfirstyear) else (.)} %>%    
    {if(!is.na(mylastyear))  filter(., year <= mylastyear) else (.)} %>%   
    
    group_by(assessyear, period) %>% 
    summarise(
      minyear = min(year),
      maxyear = max(year)
    )

  # units
  u <-
    d %>% 
    distinct(perfstat, unit2)
  
  # reference points
  rp1 <-
    rp %>% 
    filter(assess %in% myassess) %>% 
    filter(assessyear %in% myassessyear) %>% 
    distinct(perfstat, blim, bpa, msybtrigger, fmsy, flow) 
  
  # View(rp)
  
  # firstyear pblim
  p1 <-
    df %>% 
    filter(perfstat   %in% c("pblim")) %>%
    filter(stock      %in% mystock) %>% 
    filter(assess     %in% myassess) %>% 
    filter(assessyear %in% myassessyear) %>% 
    filter(method     %in% mymethod) %>% 
    filter(om         %in% myom) %>% 
    filter(mp         %in% mymp) %>% 
    filter(ftgt       %in% myftgt) %>% 
    filter(!period    %in% c("HI")) %>%
    filter(mean       < 0.05) %>% 
    
    group_by(perfstat, stock, assess, assessyear, method, om, mp, ftgt) %>% 
    summarise(mean = min(year, na.rm=TRUE))
    
    
  # recoverylines
  p2 <-
    df %>%
    filter(perfstat %in% c("firstyearrebuildtobpa")) %>%
    filter(stock      %in% mystock) %>% 
    filter(assess     %in% myassess) %>% 
    filter(assessyear %in% myassessyear) %>% 
    filter(method     %in% mymethod) %>% 
    filter(om         %in% myom) %>% 
    filter(mp         %in% mymp) %>% 
    filter(ftgt       %in% myftgt) %>% 
    
    mutate(perfstat = case_when(
      perfstat == "firstyearrebuildtoblim" ~ "precblim",
      perfstat == "firstyearrebuildtobpa"  ~ "precbpa",
      TRUE                                 ~ perfstat
    ))
  
  # title
  t <- paste(
    toupper(mystock),
    paste(unique(myassess), collapse="-"),
    paste(unique(myassessyear), collapse="-"),
    # paste(unique(mymethod), collapse="-"),
    paste(unique(myom), collapse="-"),
    paste(unique(mymp), collapse="-"),
    sep="_"
  )
  
  plist <- list(NULL)
  i <- 1
  for (i in 1:length(myperfstat)) {
    if(myperfstat[[i]] %in% c("stock","fbar","catch","rec","tac", "iav")) {
      
      print(i)
      
      plist[[i]] <-
        d %>% 
        filter(perfstat==myperfstat[[i]]) %>% 
        ggplot(aes(x=year, y=mean, group=perfstat)) +
        theme_publication() +
        theme(plot.margin = unit(c(0,0,0,0), "cm")) +
        theme(legend.position="none") +
        theme( panel.grid.major.x = element_blank()) +
        theme(strip.text = element_text(size = 12)) +

        {if (i == length(myperfstat)) 
          theme(axis.text.x = element_text(angle = 90, vjust=0.5)) else 
            theme(axis.text.x = element_blank()) } +
        
        # historical ribbon
        geom_ribbon(data=filter(h, !is.na(mean), perfstat==myperfstat[[i]]),
                    aes(ymin=lower, ymax=upper, group=iter), alpha=0.4, fill="gray") +
        
        # historical worms
        geom_line(data=filter(h, metric=="worm", perfstat==myperfstat[[i]]),
                  aes(x=year, y=data, group=iter), colour="darkgray", size=0.3) +
        
        # historical median
        geom_line(data=filter(h, !is.na(median), perfstat==myperfstat[[i]]), 
                  aes(x=year,y=median, group=iter), colour="black", size=1.5) +
    
        # future ribbon
        geom_ribbon(data=filter(d, !is.na(median), perfstat==myperfstat[[i]]),
                    aes(ymin=lower, ymax=upper, group=iter), alpha=0.3, fill=mycolours[[i]]) +
        
        # future worms
        geom_line(data=filter(d, metric=="worm", perfstat==myperfstat[[i]]),
                  aes(x=year, y=data, group=iter), colour=mycolours[[i]], size=0.3) +
        
        # future median
        geom_line(data=filter(d, !is.na(median), perfstat==myperfstat[[i]]),
                  aes(x=year,y=median,group=iter), colour=mycolours[[i]], size=1.5) +
        
        # vertical lines (periods)
        geom_vline(data=filter(p, period=="CU"),
                   aes(xintercept=minyear), linetype="dotted") +
        # geom_vline(data=filter(p, period=="ST"),
        #            aes(xintercept=minyear), linetype="dotted") +
        # geom_vline(data=filter(p, period=="MT"),
        #            aes(xintercept=minyear), linetype="dotted") +
        # geom_vline(data=filter(p, period=="LT"),
        #            aes(xintercept=minyear), linetype="dotted") +
        # geom_vline(data=filter(p, period=="LT"),
        #            aes(xintercept=maxyear), linetype="dotted") +
    
        # horizontal lines (reference points)
        # blim
        {if (myperfstat[[i]] == "stock") geom_hline(aes(yintercept=blim), linetype="dashed") } +
        {if (myperfstat[[i]] == "stock") geom_text(data=filter(rp, perfstat=="stock"), 
                                                   aes(y=blim), label="blim", x=myfirstyear,
                                                   vjust=0, hjust=0, inherit.aes = FALSE) } +
        
        # bpa
        {if (myperfstat[[i]] == "stock") geom_hline(aes(yintercept=bpa), linetype="dotted") } +
        {if (myperfstat[[i]] == "stock") geom_text(data=filter(rp, perfstat=="stock"), 
                                                   aes(y=bpa), label="bpa", x=myfirstyear, 
                                                   vjust=0, hjust=0, inherit.aes = FALSE) } +
        
        # ftgt
        {if (myperfstat[[i]] == "fbar") geom_hline(aes(yintercept=as.numeric(ftgt)), linetype="dashed") } +
        {if (myperfstat[[i]] == "fbar") geom_text(aes(y=as.numeric(ftgt)), label="ftgt", x=myfirstyear, 
                                                   vjust=0, hjust=0, inherit.aes = FALSE) } +

        # flow
        {if (myperfstat[[i]] == "fbar") geom_hline(aes(yintercept=as.numeric(ftgt)/5), linetype="dashed") } +
        {if (myperfstat[[i]] == "fbar") geom_text(aes(y=as.numeric(ftgt)/5), label="flow", x=myfirstyear, 
                                                   vjust=0, hjust=0, inherit.aes = FALSE) } +
        
        # fmsy
        {if (myperfstat[[i]] == "fbar") geom_hline(data=filter(rp1, perfstat=="fbar"), 
                                                  aes(yintercept=as.numeric(fmsy)), linetype="dashed") } +
        {if (myperfstat[[i]] == "fbar") geom_text(data=filter(rp1, perfstat=="fbar"), 
                                                  aes(y=as.numeric(fmsy)), label="fmsy", x=max(d$year), 
                                                   vjust=0, hjust=1, inherit.aes = FALSE) } +
        expand_limits(y=0) +
        
        scale_x_continuous(breaks=seq(min(h$year), max(d$year), 5)) +
        
        labs(x="", y = filter(u, perfstat == myperfstat[[i]])$unit2) +
        {if(i==1 & showtitle==TRUE) labs(title=t)} +
        
        facet_wrap( ~ perfstat)
      
  
    } else if (myperfstat[i] %in% c("pblim")) {
      
      print(i)
           
      plist[[i]] <-
        ggplot(data=filter(d, perfstat==myperfstat[[i]]), 
                   aes(x=year, y=mean, group=perfstat)) +
        theme_publication() +
        theme(plot.margin = unit(c(0,0,0,0), "cm")) +
        theme(legend.position="none") +
        theme( panel.grid.major.x = element_blank()) +
        theme(strip.text = element_text(size = 12)) +
        {if (i == length(myperfstat)) theme(axis.text.x = element_text(angle = 90, vjust=0.5)) else 
          theme(axis.text.x = element_blank()) } +
        
        geom_hline(yintercept=0.05, colour="gray", linetype="dashed") +
        
        geom_vline(data=p1, aes(xintercept=mean), colour=mycolours[[i]], linetype="dashed") +
        geom_text(data= p1,
                  aes(x=mean, label=ac(mean)),
                  y=0.05, vjust=0, hjust=0, nudge_x = 0.2, colour=mycolours[[i]]) +
        
        geom_line(data=filter(h, perfstat == myperfstat[[i]]), colour="gray", size=1) +
        geom_line(colour=mycolours[[i]], size=1) +
        
        # vertical lines (periods)
        geom_vline(data=filter(p, period=="CU"),
                   aes(xintercept=minyear), linetype="dotted") +
        # geom_vline(data=filter(p, period=="ST"),
        #            aes(xintercept=minyear), linetype="dotted") +
        # geom_vline(data=filter(p, period=="MT"),
        #            aes(xintercept=minyear), linetype="dotted") +
        # geom_vline(data=filter(p, period=="LT"),
        #            aes(xintercept=minyear), linetype="dotted") +
        # geom_vline(data=filter(p, period=="LT"),
        #            aes(xintercept=maxyear), linetype="dotted") +
    
        expand_limits(y=0) +
        scale_x_continuous(breaks=seq(min(h$year), max(d$year), 5)) +
        
        labs(x="", y = filter(u, perfstat == myperfstat[[i]])$unit2) +
        {if(i==1 & showtitle==TRUE) labs(title=t)} +
        
        facet_wrap( ~ perfstat)
  
    } else if (myperfstat[i] %in% c("precbpa")) {
      
      print(i)
           
      plist[[i]] <-
        ggplot(data=filter(d, perfstat==myperfstat[[i]]), 
                   aes(x=year, y=mean, group=perfstat)) +
        theme_publication() +
        theme(plot.margin = unit(c(0,0,0,0), "cm")) +
        theme(legend.position="none") +
        theme( panel.grid.major.x = element_blank()) +
        theme(strip.text = element_text(size = 12)) +
        
        {if (i == length(myperfstat)) {
          theme(axis.text.x = element_text(angle = 90, vjust=0.5))
        } else {
          theme(axis.text.x = element_blank())
        } } +
        
        geom_vline(data=p2, aes(xintercept=mean), colour=mycolours[[i]], linetype="dashed") +
        
        geom_hline(yintercept=0.5, colour="gray", linetype="dashed") +
        
        geom_vline(data=filter(p, period=="CU"),
                   aes(xintercept=minyear), linetype="dotted") +
        
        geom_text(data=filter(p2, perfstat %in% c("precbpa")),
                  aes(x=mean, label=ac(mean)),
                  y=0.0, vjust=0, hjust=0, nudge_x = 0.2, colour=mycolours[[i]]) +
    
        geom_line(colour=mycolours[[i]], size=1) +

        expand_limits(y=0, x=min(p$minyear)) +
        scale_x_continuous(breaks=seq(min(h$year), max(d$year), 5)) +
        
        labs(x="", y = filter(u, perfstat == myperfstat[[i]])$unit2) +
        # {if(i==1 & showtitle==TRUE) labs(title=t)} +
        
        facet_wrap( ~ perfstat)

    } # end of if statements
  
  } # end of for loop

print(patchwork::wrap_plots(plist) + 
  patchwork::plot_layout(ncol = ncolumns) +
  patchwork::plot_annotation() & theme(plot.tag = element_text(size = 8)))

ggsave(file = file.path(Res.dir,paste0(t," scenarioplot.png")),
       device="png", width = 30, height = 20, units = "cm")

  
} # end of function


# mystock      =  "WHOM"
# myassess     = c("SS3", "SAM")
# myassessyear = c("2022")
# mymethod     = "EQSIM"
# myom         = c("OM2.3", "OM2.5")
# mymp         = c("MP5.23")
# myperfstat   = c("catch","stock", "fbar", "pblim", "precbpa", "iav")
# myftgt       = c(0.0, 0.05, 0.074, 0.10, 0.15)
# myfirstyear  = 2023
# mylastyear   = 2030

generatetable <- function(
                    mystock      =  "WHOM",
                    myassess     = c("SS3", "SAM"),
                    myassessyear = c("2022"),
                    mymethod     = "EQSIM",
                    myom         = c("OM2.3", "OM2.5"),
                    mymp         = c("MP5.23"),
                    myperfstat   = c("catch","stock", "fbar", "iav", "pblim", "precbpa"),
                    myftgt       = c(0.0, 0.05, 0.074, 0.10, 0.15),
                    myfirstyear  = 2023,
                    mylastyear   = 2030) {
  
  t <-
    df %>% 
    filter(mp %in% mymp) %>% 
    filter(assess %in% myassess) %>%  
    filter(assessyear %in% myassessyear) %>%  
    filter(perfstat %in% myperfstat) %>%
    filter(ftgt %in% myftgt) %>% 
    filter(year >= myfirstyear) %>%
    filter(year <= mylastyear) %>% 
    filter(!is.na(median) | !is.na(mean)) %>%
    mutate(data = ifelse(is.na(median), mean, median)) %>% 
    mutate(data = ifelse(perfstat %in% c("catch", "stock"), 1000*data, data))

  d1 <- data.frame(
          perfstat = c("catch", "stock","fbar", "iav", "pblim", "precbpa"), 
          round    = c(      0,       0,     2,     2,        2,        2))

  for(tst in myperfstat) {
    t %>% 
      filter(perfstat==tst) %>% 
      reshape2::dcast(assess+assessyear+perfstat+ftgt ~ year, value.var="data", sum, margins=NA) %>% 
      ungroup() %>% 
      pander::pandoc.table(., 
                   style        = "simple",
                   split.tables = 200, 
                   split.cells  = c(rep(7,10)),
                   justify      = "right",
                   missing      = " ",
                   big.mark     = ',', 
                   round        = d1[d1$perfstat==tst,2])
    
  }
  

} # end of generatetable




```

*Working document for NPWG/PELAC 2022*


Martin Pastoors, `r format(Sys.time(), '%d/%m/%Y')`

&nbsp;  

**Abstract**

During 2020 and 2021, an evaluation of a potential rebuilding plan for Western Horse mackerel was carried out and reviewed by ICES. ICES concluded that "the evaluated rebuilding plan as proposed by PELAC shows potential to reach the specified target (three consecutive years > Bpa) within the time frame specified in the plan (< ten years) and is considered to be precautionary in the long term.". Although the PELAC rebuilding plan option was subsequently included in the ICES options table for western horse mackerel in 2021, the TAC setting for 2022 was based on the ICES MSY advice and not on the proposed rebuilding plan. 

At WGWIDE 2022 group, the assessment of western horse mackerel indicated that the stock was now below Blim and the stock could not be rebuilt to Blim within the two years of the short term forecast. Therefore ICES recommended a zero catch for Western horse mackerel in 2023. ICES also included the PELAC rebuilding plan option in the forecast table (15513 tonnes in 2023). 

This document provides an update of the MSE evaluations from 2020-2021 using the results of WGWIDE 2022 and either the SS3 model or the SAM model and using the preferred rebuidling plan option: double breakpoint rule (target fishing mortality at Fmsy=0.074, breakpoint at MSY Btrigger and straight decline in F to 20% of Ftarget at Blim. Below Blim continued fishing at F = 0.2 * Ftarget). Results are presented in this working document in comparison with the results obtained presented in WKWHMRP 2021 for the preferred rebuilding plan options. All simulations based on the WGWIDE 2022 data are based on the assumption that the TAC in 2022 will be taken and that the TAC in 2023 will be set on the basis of the rebuilding plan (15513 tonnes).

The code used to carry out the MSE process, was the same code as used in 2020 but updated to take into account the assessments as carried out during WGWIDE 2022 (see: https://github.com/martinpastoors/homrebuild).  

The rebuilding potential from the SS3 assessment of WGWIDE 2022 is substantially slower compared to the rebuilding potential that was estimated from the 2020 assessment. This is likely caused by the stock being estimated to be lower over a long number of years and in recent years also below Blim. Rebuilding the stock to MSY Btrigger for at least 3 years in a row and for at least 50% of the simulations is now expected to take until 2035 while in the previous evaluation this was 2028. 

The estimated starting conditions and rebuilding period for the SAM assessment of WGWIDE 2022 is also a bit longer compared to the 2020 assessment. The new evaluation estimates rebuilding in year 2028 while in the 2020 evaluation this was in 2025. 

In both evaluations (SS3 and SAM) the target fishing mortality of the rebuilding plan (0.074) are not achieved in the simulations, due to the constraints on TAC variation that applies when the stock is above MSY Btrigger. The low starting TAC for the simulations (15513 tonnes for 2023) also affect this behavior. 

\newpage

# Introduction

During 2020, a group of experts associated with the Pelagic Advisory Council (PELAC) developed an evaluation of a potential rebuilding plan for Western Horse mackerel (Pastoors et al., 2020). The development of the rebuilding plan was instigated by the PELAC to address the SSB estimates of the stock that were hovering just above Blim following the reference points interbenchmark in 2019 (ICES 2019) and the WGWIDE assessment in 2020 (ICES 2020). The evaluation was submitted to ICES for scientific review by the European Commission. ICES then convened the WKWHMRP to carry out the scientific review (ICES 2021a) which resulted in an ICES advice in April 2021 (ICES 2021a). ICES concluded that "the evaluated rebuilding plan as proposed by PELAC shows potential to reach the specified target (three consecutive years > Bpa) within the time frame specified in the plan (< ten years) and is considered to be precautionary in the long term.". 

During the annual advice cycle in 2021, the PELAC rebuilding plan option was included in the options table for western horse mackerel, although the headline advice was still based on the ICES MSY approach, because the rebuilding plan had not been adopted by all relevant member states. During the TAC setting discussions between EU and UK, the TAC for Western horse mackerel for 2022 was set on the basis of the ICES MSY advice and not on the proposed rebuilding plan. 

At the WGWIDE 2022 group, the assessment of western horse mackerel indicated that the stock was now below Blim (due to a retrospective downward revision of the assessment over many years). The stock could not be rebuilt to Blim within the two years of the short term forecast, and therefore ICES recommended a zero catch for Western horse mackerel in 2023. ICES also included the PELAC rebuilding plan option in the forecast table, which would result in a TAC of 15513 tonnes in 2023. 

A comparison of the PELAC recommended rebuilding plan and the ICES MSY Advice rule is presented in the figure below. Here the ICES MSY advice rule has been interpreted in a strict sense, meaning that the recommended F is decreased from a relatively high value just above Blim to zero when the stock is below Blim. In contrast, the PELAC recommended rebuilding plan has a much stronger decrease in F between MSY Btrigger and Blim, and a continued, low fishing mortality when the stock is below Blim. In addition, the rebuilding plan as a 20% constraint on TAC changes when the stock is above MSY Btrigger.   


```{r echo=FALSE, fig.align="center", fig.asp=0.4, message=FALSE, warning=FALSE}

readxl::read_excel("C:/DATA/GIT/homrebuild/data/hcr_types v2.xlsx")  %>% 
  mutate(scenario=factor(scenario, levels=c("ICES AR", "Double BP + IAV"))) %>% 
  
  ggplot(aes(x=ssb,y=f, group=scenario)) +
  theme_publication() +
  theme(panel.grid.major = element_blank()) +
  theme(legend.position = "none") +
  geom_vline(xintercept=834, colour = "darkgray", linetype="dotted", size=0.6) +
  geom_vline(xintercept=1168, colour = "darkgray", linetype="dotted", size=0.6) +
  geom_hline(yintercept=0.074, colour = "darkgray", linetype="dotted", size=0.6) +
  geom_ribbon(aes(fill=scenario, ymin=flow, ymax=fupp), alpha=0.3) +
  geom_line(aes(colour=scenario), size=1) +
  # scale_linetype(c("solid","dashed")) +
  expand_limits(y=0) +
  guides(linetype=FALSE) +
  facet_wrap(~scenario)

```

Following the release of the ICES advice, and in consultation with several members of the PELAC, a part of the MSE analysis that was done in the initial evaluation of the rebuilding plan, has been repeated in the current working documents. It provides an update of the MSE evaluations using either the SS3 model (the model used by WGWIDE for western horse mackerel) or the SAM model (used as exploratory assessment in WGWIDE) and using the preferred rebuidling plan option: double breakpoint rule (target fishing mortality at Fmsy=0.074, breakpoint at MSY Btrigger and straight decline in F to 20% of Ftarget at Blim. Below Blim continued fishing at F = 0.2 * Ftarget). Results are presented in this working document in comparison with the results obtained presented in WKWHMRP 2021 for the preferred rebuilding plan options. 

The code used to carry out the MSE process, was the same code as used in 2020 but updated to take into account the assessments as carried out during WGWIDE 2022 (for code, see: https://github.com/martinpastoors/homrebuild). No assessment or forecast is conducted during the simulations i.e. this is a ‘short-cut’ MSE approach. The parameterisation of uncertainty in observed SSB and F are incorporated into the simulations following the methodology developed for EqSim whereby SSB and FBar outputs from historic assessments and short term forecasts are compared to the most recent assessment and described via a pair of parameters (CV and autocorrelation) for each of SSB and FBar. These parameters are then used to generate an error and derived “observed” values of SSB and Fbar for the management procedure (Pastoors 2020). 

All simulations based on the WGWIDE 2022 data are based on the assumption that the TAC in 2022 will be taken and that the TAC in 2023 will be set on the basis of the rebuilding plan (15513 tonnes). 
 

# Results

Results are presented for MSEs that are either based on the WGWIDE 2020 or WGWIDE 2022 assessments of western horse mackerel, and using either the SS3 assessment model (which is the basis for the ICES advice) or the SAM model (which is used as exploratory assessment in WGWIDE). The results based on WGWIDE 2020 are identical to the results that were presented to WKWHMRP 2021 (ICES 2021a). The results based on WGWIDE 2022 and new and presented here for the first time. The target fishing mortality in the simulations is 0.074 (=Fmsy). 

Figures 1-4 contain the historical patterns and the simulated results for SSB, fishing mortality, recruitment and catch. The development of stock in both the SS3 and SAM evaluations indicate that the stock is expected to increase from the lowest biomass around 2018-2018.

The scenarios presented all have a target fishing mortality of 0.074 (Fmsy). However, this is not achieved, especially in the cases when the 2022 assessments are used. This is because the simulations start with a TAC of 15 kT, and subsequently, when the stock is estimated above MSY Btrigger, the change in TAC is limited to 20%. This means that the maximum increase in TAC is limiting the realised fishing mortality. 


```{r echo=FALSE, fig.align="center", fig.asp=0.9, message=FALSE, warning=FALSE}

plotvar(mystock      = "WHOM",
        myassess     = c("SS3", "SAM"),
        myassessyear = c("2020", "2022"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.3", "OM2.5"),
        mymp         = c("MP5.23"),
        myftgt       = c(0.074),
        myperfstat   = c("stock"),
        mycolour     = "red",
        myyintercept = "blim",
        myvalue      = "median",
        myfacets     = c("assess", "assessyear"),
        myfirstyear  = 2000,
        scale_y_free = TRUE)


```

_fig 1 Simulation summaries of SSB using MP5.23 (double breakpoint F with 20% IAV constraint above Btrigger) at F=0.074 (Fmsy) from either WGWIDE 2020 or WGWIDE 2022 and either SS3 (default assessment) or SAM (exploratory assessment)._

```{r echo=FALSE, fig.align="center", fig.asp=0.9, message=FALSE, warning=FALSE}

plotvar(mystock      = "WHOM",
        myassess     = c("SS3", "SAM"),
        myassessyear = c("2020", "2022"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.3", "OM2.5"),
        mymp         = c("MP5.23"),
        myftgt       = c(0.074),
        myperfstat   = c("fbar"),
        mycolour     = "blue",
        myyintercept = "blim",
        myvalue      = "median",
        myfacets     = c("assess", "assessyear"),
        myfirstyear  = 2000,
        scale_y_free = TRUE)


```

_fig 2 Simulation summaries of fishing mortality using MP5.23 (double breakpoint F with 20% IAV constraint above Btrigger) at F=0.074 (Fmsy) from either WGWIDE 2020 or WGWIDE 2022 and either SS3 (default assessment) or SAM (exploratory assessment)._

\newpage

```{r echo=FALSE, fig.align="center", fig.asp=0.9, message=FALSE, warning=FALSE}

plotvar(mystock      = "WHOM",
        myassess     = c("SS3", "SAM"),
        myassessyear = c("2020", "2022"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.3", "OM2.5"),
        mymp         = c("MP5.23"),
        myftgt       = c(0.074),
        myperfstat   = c("rec"),
        mycolour     = "darkgreen",
        myyintercept = "blim",
        myvalue      = "median",
        myfacets     = c("assess", "assessyear"),
        myfirstyear  = 2000,
        scale_y_free = TRUE)


```
_fig 3 Simulation summaries of recruitment using MP5.23 (double breakpoint F with 20% IAV constraint above Btrigger) at F=0.074 (Fmsy) from either WGWIDE 2020 or WGWIDE 2022 and either SS3 (default assessment) or SAM (exploratory assessment)._


```{r echo=FALSE, fig.align="center", fig.asp=0.9, message=FALSE, warning=FALSE}

plotvar(mystock      = "WHOM",
        myassess     = c("SS3", "SAM"),
        myassessyear = c("2020", "2022"),
        mymethod     = c("EQSIM"),
        myom         = c("OM2.3", "OM2.5"),
        mymp         = c("MP5.23"),
        myftgt       = c(0.074),
        myperfstat   = c("catch"),
        mycolour     = "darkred",
        myyintercept = "blim",
        myvalue      = "median",
        myfacets     = c("assess", "assessyear"),
        myfirstyear  = 2000,
        scale_y_free = TRUE)


```

_fig 4 Simulation summaries of catch using MP5.23 (double breakpoint F with 20% IAV constraint above Btrigger) at F=0.074 (Fmsy) from either WGWIDE 2020 or WGWIDE 2022 and either SS3 (default assessment) or SAM (exploratory assessment)._

\newpage

**Probability of being below Blim**

The probability of being below Blim is an important criteria for the assessment of the precautionary nature of management actions and long term plans. When the probability of being below Blim is less than 5%, this is considered in accordance with the precautionary approach. With the 2022 assessments, the probability of being below Blim is decreasing. The expected year when the 5% probability would be achieved with the SS3 assessment is substantially futher into the future.

```{r echo=FALSE, fig.align="center", fig.asp=0.9, message=FALSE, warning=FALSE}

plotpblim (  mystock      = "WHOM",
             myassess     = c("SS3", "SAM"),
             myassessyear = c("2020", "2022"),
             mymethod     = c("EQSIM"),
             myom         = c("OM2.3", "OM2.5"),
             mymp         = c("MP5.23"),
             myftgt       = c(0.074),
             myfacets     =  c("assess", "assessyear"),
             myfirstyear  = 2005,
             mycolour     = "red")


```

_fig 5 Simulation summaries of probability of being below blim using MP5.23 (double breakpoint F with 20% IAV constraint above Btrigger) at F=0.074 (Fmsy) from either WGWIDE 2020 or WGWIDE 2022 and either SS3 (default assessment) or SAM (exploratory assessment)._

\newpage

**Probability of rebuilding to MSY Btrigger (Bpa)**

The probability of rebuilding to MSY Btrigger (Bpa) is defined as 50% of the simulations being above MSY Btrigger for at least 3 subsequent years. 

For the SS3 assessment, the year of rebuilding to Bpa would extended from 2018 to 2035, while for the SAM assessment this is from 2025 to 2028. The difference between SS3 and SAM is generally that the SS3 assessment gives a somewhat more pessimistic stock development trajectory compared to the SAM assessment (note that the SAM assessment does not estimate the stock to be below Blim in 2022). 

```{r echo=FALSE, fig.align="center", fig.asp=0.9, message=FALSE, warning=FALSE}


plotrecovery(mystock      = "WHOM",
             myassess     = c("SS3", "SAM"),
             myassessyear = c("2020", "2022"),
             mymethod     = c("EQSIM"),
             myom         = c("OM2.3", "OM2.5"),
             mymp         = c("MP5.23"),
             myftgt       = c(0.074),
             myfacets     =  c("assess", "assessyear"),
             myfirstyear  = 2005,
             mylastyear   = max(df$year),
             myintercept  = 0.5)

```

_fig 6 Simulation summaries of recovery to blim or bpa (at least 50% above reference point for 3 subsequent years) using MP5.23 (double breakpoint F with 20% IAV constraint above Btrigger) at F=0.074 (Fmsy) from either WGWIDE 2020 or WGWIDE 2022 and either SS3 (default assessment) or SAM (exploratory assessment)._

\newpage

**Summary**

A summary of the SS3 run with target F = 0.074 is shown in Figure 7 and Table 1. 

```{r echo=FALSE, fig.align="center", fig.asp=0.9, message=FALSE, warning=FALSE}

plotscenario (
    mystock      =  "WHOM",
    myassess     = "SS3",
    myassessyear = c("2022"),
    mymethod     = "EQSIM",
    myom         = c("OM2.3"),
    mymp         = c("MP5.23"),
    myftgt       = c(0.074),
    myfacets     = c("perfstat"),
    myfirstyear  = 2000,
    mylastyear   = as.numeric(NA),
    myperfstat   = c("stock","fbar", "rec", "catch", "iav", "pblim", "precbpa"),
    mycolours    = c("blue","red","darkgreen", "brown","blue","red", "black"),
    ncolumns     = 2)

```

_fig 7 Simulation summaries of several relevant quantities using MP5.23 (double breakpoint F with 20% IAV constraint above Btrigger) at F=0.074 (Fmsy) from WGWIDE 2022 and SS3 (default assessment)._

\newpage

**Table 1: summary of SS3 2022 simulation results at Ftgt=0.074**

```{r echo=FALSE, fig.align="center", fig.asp=0.55, message=FALSE, warning=FALSE}


generatetable (mystock      =  "WHOM",
               myassess     = c("SS3"),
               myassessyear = c("2022"),
               mymethod     = "EQSIM",
               myom         = c("OM2.3"),
               mymp         = c("MP5.23"),
               myperfstat   = c("catch","stock", "fbar", "pblim", "precbpa"),
               myftgt       = c(0.001, 0.05, 0.074, 0.10, 0.15),
               myfirstyear  = 2023,
               mylastyear   = 2030) 

```


\newpage

A summary of the SAM run with target F = 0.074 is shown in Figure 8 and Table 2. 

```{r echo=FALSE, fig.align="center", fig.asp=0.9, message=FALSE, warning=FALSE}

plotscenario (
    mystock      =  "WHOM",
    myassess     = "SAM",
    myassessyear = c("2022"),
    mymethod     = "EQSIM",
    myom         = c("OM2.5"),
    mymp         = c("MP5.23"),
    myftgt       = c(0.074),
    myfacets     = c("perfstat"),
    myfirstyear  = 2000,
    mylastyear   = as.numeric(NA),
    myperfstat   = c("stock","fbar", "rec", "catch", "iav", "pblim", "precbpa"),
    mycolours    = c("blue","red","darkgreen", "brown","blue","red", "black"),
    ncolumns     = 2)

```

_fig 8 Simulation summaries of several relevant quantities using MP5.23 (double breakpoint F with 20% IAV constraint above Btrigger) at F=0.074 (Fmsy) from WGWIDE 2022 and SAM (exploratory assessment)._

\newpage

**Table 2: summary of SAM 2022 simulation results at Ftgt=0.074**

```{r echo=FALSE, fig.align="center", fig.asp=0.55, message=FALSE, warning=FALSE}


generatetable (mystock      =  "WHOM",
               myassess     = c("SAM"),
               myassessyear = c("2022"),
               mymethod     = "EQSIM",
               myom         = c("OM2.5"),
               mymp         = c("MP5.23"),
               myperfstat   = c("catch","stock", "fbar", "pblim", "precbpa"),
               myftgt       = c(0.001, 0.05, 0.074, 0.10, 0.15),
               myfirstyear  = 2023,
               mylastyear   = 2030) 


```


# Discussion and conclusions

During 2020 and 2021, an evaluation of a potential rebuilding plan for Western Horse mackerel was carried out and reviewed by ICES. ICES concluded that "the evaluated rebuilding plan as proposed by PELAC shows potential to reach the specified target (three consecutive years > Bpa) within the time frame specified in the plan (< ten years) and is considered to be precautionary in the long term.". Although the PELAC rebuilding plan option was subsequently included in the ICES options table for western horse mackerel in 2021, the TAC setting for 2022 was based on the ICES MSY advice and not on the proposed rebuilding plan. 

At WGWIDE 2022 group, the assessment of western horse mackerel indicated that the stock was now below Blim and the stock could not be rebuilt to Blim within the two years of the short term forecast. Therefore ICES recommended a zero catch for Western horse mackerel in 2023. ICES also included the PELAC rebuilding plan option in the forecast table (15513 tonnes in 2023). 

This document provides an update of the MSE evaluations from 2020-2021 using the results of WGWIDE 2022 and either the SS3 model or the SAM model and using the preferred rebuidling plan option: double breakpoint rule (target fishing mortality at Fmsy=0.074, breakpoint at MSY Btrigger and straight decline in F to 20% of Ftarget at Blim. Below Blim continued fishing at F = 0.2 * Ftarget). Results are presented in this working document in comparison with the results obtained presented in WKWHMRP 2021 for the preferred rebuilding plan options. All simulations based on the WGWIDE 2022 data are based on the assumption that the TAC in 2022 will be taken and that the TAC in 2023 will be set on the basis of the rebuilding plan (15513 tonnes).

The code used to carry out the MSE process, was the same code as used in 2020 but updated to take into account the assessments as carried out during WGWIDE 2022 (see: https://github.com/martinpastoors/homrebuild).  

The rebuilding potential from the SS3 assessment of WGWIDE 2022 is substantially slower compared to the rebuilding potential that was estimated from the 2020 assessment. Rebuilding the stock to MSY Btrigger for at least 3 years in a row and for at least 50% of the simulations is now expected to take until 2035 while in the previous evaluation this was 2028. 

The estimated starting conditions and rebuilding period for the SAM assessment of WGWIDE 2022 is also a bit longer compared to the 2020 assessment. The new evaluation estimates rebuilding in year 2028 while in the 2020 evaluation this was in 2025. 

The reason that the simulations using the 2022 assessments are different from the simulations from the 2020 assessment is largely in relation to the biomass at the start of the simulation period relative to the reference points Blim and Bpa. Figures 9 and 10 show the biomass trends from the assessments and simulated periods relative to Blim and Bpa. The 2022 SAM assessment estimates the SSB to be between Blim and Bpa at the start of the simulation period, whereas the SS3 assessment estimates the SSB to be below Blim. 

```{r echo=FALSE, fig.align="center", fig.asp=0.9, message=FALSE, warning=FALSE}

t <-
  df %>% 
  filter(perfstat=="stock", ftgt==0.074) %>%
  filter(year >= 2000) %>% 
  filter(mp == "MP5.23") %>% 
  mutate(assess = factor(assess, levels=c("SS3", "SAM"))) %>% 
  drop_na(median) %>% 
  dplyr::select(stock, assess, assessyear, period, year, lower, median, upper, blim, bpa) %>% 
  tidyr::pivot_longer(names_to = "estimate", values_to = "data", c("lower", "median", "upper")) %>% 
  mutate(rel_blim = data/blim) %>% 
  mutate(rel_bpa  = data/bpa) %>% 
  dplyr::select(-bpa, -blim, -data) %>% 
  tidyr::pivot_longer(names_to = "variable", values_to = "data", c("rel_blim", "rel_bpa")) %>% 
  tidyr::pivot_wider(names_from = estimate, values_from = data)
  # View()

p <-
  t %>% 
  distinct(assessyear, period, year) %>% 
  
  group_by(assessyear, period) %>% 
  summarise(
    minyear = min(year),
    maxyear = max(year)
  )
  

t %>% 
  filter(variable == "rel_blim") %>% 
  
  ggplot(aes(x=year, y=median)) +
  theme_publication() +
  theme( panel.grid.major.x = element_blank()) +

  geom_ribbon(aes(ymin=lower, ymax=upper), alpha=0.4, fill="gray") +
  geom_line(aes(y=lower), colour="darkgray") +
  geom_line(aes(y=upper), colour="darkgray") +
  
  geom_line(size=1) +
  geom_hline(yintercept=1, colour="red", linetype="dashed") +
  
  # vertical lines (periods)
  geom_vline(data=filter(p, period=="CU"),
             aes(xintercept=minyear), linetype="dotted") +
  geom_vline(data=filter(p, period=="ST"),
             aes(xintercept=minyear), linetype="dotted") +
  geom_vline(data=filter(p, period=="MT"),
             aes(xintercept=minyear), linetype="dotted") +
  geom_vline(data=filter(p, period=="LT"),
             aes(xintercept=minyear), linetype="dotted") +
  geom_vline(data=filter(p, period=="LT"),
             aes(xintercept=maxyear), linetype="dotted") +
    
  labs(y="SSB/Blim", x="", title="SSB relative to Blim") +
  facet_grid(assess ~ assessyear)

```

_fig 9 Simulation summaries of SSB relative to Blim using MP5.23 (double breakpoint F with 20% IAV constraint above Btrigger) at F=0.074 (Fmsy) from SS3 and SAM assessments._


```{r echo=FALSE, fig.align="center", fig.asp=0.9, message=FALSE, warning=FALSE}

# data processing in previous chunk

t %>% 
  filter(variable == "rel_bpa") %>% 
  
  ggplot(aes(x=year, y=median)) +
  theme_publication() +
  theme( panel.grid.major.x = element_blank()) +
  
  geom_ribbon(aes(ymin=lower, ymax=upper), alpha=0.4, fill="gray") +
  geom_line(aes(y=lower), colour="darkgray") +
  geom_line(aes(y=upper), colour="darkgray") +
  
  geom_line(size=1) +
  
  # vertical lines (periods)
  geom_vline(data=filter(p, period=="CU"),
             aes(xintercept=minyear), linetype="dotted") +
  geom_vline(data=filter(p, period=="ST"),
             aes(xintercept=minyear), linetype="dotted") +
  geom_vline(data=filter(p, period=="MT"),
             aes(xintercept=minyear), linetype="dotted") +
  geom_vline(data=filter(p, period=="LT"),
             aes(xintercept=minyear), linetype="dotted") +
  geom_vline(data=filter(p, period=="LT"),
             aes(xintercept=maxyear), linetype="dotted") +
  
  geom_hline(yintercept=1, colour="red", linetype="dashed") +
  labs(y="SSB/Bpa", x="", title="SSB relative to Bpa") +
  facet_grid(assess ~ assessyear)


```

_fig 10 Simulation summaries of SSB relative to Bpa using MP5.23 (double breakpoint F with 20% IAV constraint above Btrigger) at F=0.074 (Fmsy) from SS3 and SAM assessments._

\newpage

Figure 11 shows the fishing mortality from the assessment and forward simulation (in red) against the fishing mortality that would result from the application of the F-rule within the HCR (in gray). 

The performance of the Double breakpoint HCR rule with contraint on interannual variation in TACs (20%) above MSY Btrigger is strongly influenced by the starting conditions (TAC and SSB in the first year) of the simulations. In this evaluation, we assumed that the TAC for 2023 would be set at 15513 tonnes, in line with the application of the rebuilding plan and the ICES forecast table. Below MSY Btrigger there are no constraints on the changes in TAC, but the target F is sharply reduced between MSY Btrigger and Blim in accordance with the HCR. If the stock were to be estimated above MSY Btrigger, the constraint on changes in TAC kicks in. 

For the SAM simulations, the median estimates of SSB are between Blim and MSY Btrigger at the start of the simulation (Figures 9 and 10). Therefore, the fishing mortalities during the rebuilding period are mostly determined by the F-rule. Once the SSB is estimated above MSY Btrigger, the TAC is not just determined by the F-rule but also by the constraint on TAC (increase). Because, SSB reaches MSY Btrigger relatively quickly in the SAM simulations, this causes the realised F to be substantially below the target F, simply because the catch is constrained by the TAC constraint.   

For the SS3 simulations, the stock is estimated to remain below Blim for some time, leading to the application of Flow (Ftarget/5) for those years. When the stock is between Blim and MSY Btrigger, the realised F is slowly increasing with the slow rebuilding of the stock. As with the SAM simulations, the SS3 simulations lead to a lower realised fishing mortality compared to the target F. 

```{r echo=FALSE, fig.align="center", fig.asp=0.9, message=FALSE, warning=FALSE}

t <-
  df %>% 
  filter(perfstat=="stock", ftgt==0.074, ) %>%
  filter(year >= 2000) %>% 
  filter(mp == "MP5.23") %>% 
  mutate(assess = factor(assess, levels=c("SS3", "SAM"))) %>% 
  drop_na(median) %>% 
  dplyr::select(stock, assess, assessyear, ftgt, period, year, lower, median, upper, blim, bpa) %>% 
  tidyr::pivot_longer(names_to = "estimate", values_to = "data", c("lower", "median", "upper")) %>% 
  mutate(ftgt = as.numeric(ftgt)) %>% 
  mutate(fhcr = ftgt/5) %>% 
  mutate(fhcr = ifelse(data >  blim & data <= bpa, 
                       ftgt/5 + (data - blim)/(bpa-blim) * (ftgt-ftgt/5), fhcr)) %>% 
  mutate(fhcr = ifelse(data >= bpa, ftgt, fhcr)) %>% 
  dplyr::select(-bpa, -blim, -data) %>% 
  tidyr::pivot_wider(names_from = estimate, values_from = fhcr)
  # View()

t2 <-
  df %>% 
  mutate(assess = factor(assess, levels=c("SS3", "SAM"))) %>% 
  filter(perfstat=="fbar", ftgt==0.074) %>%
  filter(year >= 2000) %>% 
  filter(mp == "MP5.23") %>% 
  drop_na(median) 

p <-
  t %>% 
  distinct(assessyear, period, year) %>% 
  
  group_by(assessyear, period) %>% 
  summarise(
    minyear = min(year),
    maxyear = max(year)
  )
  

t %>% 
  ggplot(aes(x=year, y=median)) +
  theme_publication() +
  theme( panel.grid.major.x = element_blank()) +

  geom_ribbon(aes(ymin=lower, ymax=upper), alpha=0.4, fill="gray") +
  geom_line(aes(y=lower), colour="darkgray") +
  geom_line(aes(y=upper), colour="darkgray") +
  
  geom_line(size=1) +

  geom_ribbon(data=t2, aes(ymin=lower, ymax=upper), alpha=0.2, fill="red") +
  geom_line(data=t2, aes(x=year, y=median), colour="red") +
  
  # vertical lines (periods)
  geom_vline(data=filter(p, period=="CU"),
             aes(xintercept=minyear), linetype="dotted") +
  geom_vline(data=filter(p, period=="ST"),
             aes(xintercept=minyear), linetype="dotted") +
  geom_vline(data=filter(p, period=="MT"),
             aes(xintercept=minyear), linetype="dotted") +
  geom_vline(data=filter(p, period=="LT"),
             aes(xintercept=minyear), linetype="dotted") +
  geom_vline(data=filter(p, period=="LT"),
             aes(xintercept=maxyear), linetype="dotted") +
    
  labs(y="F HCR", x="", title="F according to F-rule (gray) and realized (red)") +
  expand_limits(y=0) +
  facet_grid(assess ~ assessyear)


```


_fig 11 Simulation summaries of fishing mortality from assessment and simulation (red) and from application of the F-rule within the HCR (gray) using MP5.23 (double breakpoint F with 20% IAV constraint above Btrigger) at F=0.074 (Fmsy) from SS3 and SAM assessments._

\newpage

In order to check the behaviour of the IAV constraint in the HCR, the SAM 2022 simulation was rerun with Management Procedure 5.20 (double breakpoint HCR, without IAV constraint above MSY Btrigger, Figure 12). Results indicate that, indeed, the IAV constraint limits the increase in TAC from a low level, thereby leading to a low realised fishing mortality (and a faster increase in stock size).   

```{r echo=FALSE, fig.align="center", fig.asp=0.9, message=FALSE, warning=FALSE}

t <-
  df %>% 
  filter(perfstat=="stock", ftgt==0.074) %>%
  filter(year >= 2000) %>% 
  filter(assessyear==2022) %>% 
  mutate(assess = factor(assess, levels=c("SS3", "SAM"))) %>%
  drop_na(median) %>% 
  mutate(assessmp = paste(assess, mp, sep="_")) %>% 
  dplyr::select(stock, assess, assessyear, mp, assessmp, ftgt, period, year, lower, median, upper, blim, bpa) %>% 
  tidyr::pivot_longer(names_to = "estimate", values_to = "data", c("lower", "median", "upper")) %>% 
  mutate(ftgt = as.numeric(ftgt)) %>% 
  mutate(fhcr = ftgt/5) %>% 
  mutate(fhcr = ifelse(data >  blim & data <= bpa, 
                       ftgt/5 + (data - blim)/(bpa-blim) * (ftgt-ftgt/5), fhcr)) %>% 
  mutate(fhcr = ifelse(data >= bpa, ftgt, fhcr)) %>% 
  dplyr::select(-bpa, -blim, -data) %>% 
  tidyr::pivot_wider(names_from = estimate, values_from = fhcr)
  # View()

t2 <-
  df %>% 
  mutate(assess = factor(assess, levels=c("SS3", "SAM"))) %>%
  filter(perfstat=="fbar", ftgt==0.074) %>%
  filter(year >= 2000) %>% 
  filter(assessyear==2022) %>% 
  # filter(assess=="SAM") %>% 
  mutate(assessmp = paste(assess, mp, sep="_")) %>% 
  drop_na(median) 

p <-
  t %>% 
  distinct(assessyear, period, year) %>% 
  
  group_by(assessyear, period) %>% 
  summarise(
    minyear = min(year),
    maxyear = max(year)
  )
  
t %>% 
  ggplot(aes(x=year, y=median)) +
  theme_publication() +
  theme( panel.grid.major.x = element_blank()) +

  geom_ribbon(aes(ymin=lower, ymax=upper), alpha=0.4, fill="gray") +
  geom_line(aes(y=lower), colour="darkgray") +
  geom_line(aes(y=upper), colour="darkgray") +
  
  geom_line(size=1) +

  geom_ribbon(data=t2, aes(ymin=lower, ymax=upper), alpha=0.2, fill="red") +
  geom_line(data=t2, aes(x=year, y=median), colour="red") +
  
  # vertical lines (periods)
  geom_vline(data=filter(p, period=="CU"),
             aes(xintercept=minyear), linetype="dotted") +
  geom_vline(data=filter(p, period=="ST"),
             aes(xintercept=minyear), linetype="dotted") +
  geom_vline(data=filter(p, period=="MT"),
             aes(xintercept=minyear), linetype="dotted") +
  geom_vline(data=filter(p, period=="LT"),
             aes(xintercept=minyear), linetype="dotted") +
  geom_vline(data=filter(p, period=="LT"),
             aes(xintercept=maxyear), linetype="dotted") +
    
  labs(y="F HCR", x="", title="2022 assessment: F according to F-rule (gray) and realized (red) without (MP5.20) and with (MP5.23) TAC constraint") +
  expand_limits(y=0) +
  facet_grid(assess ~ mp)


```

_fig 12 Simulation summaries of 2022 assessments for fishing mortality from assessment and simulation (red) and from application of the F-rule within the double breakpoint HCR (gray) without IAV constraint (MP5.20) and with IAV constraint (MP5.23) at F=0.074 (Fmsy)._


# References

ICES. 2019. Interbenchmark Protocol on reference points for western horse mackerel (Trachurus trachurus) in subarea 8 and divisions 2.a, 4.a, 5.b, 6.a, 7.a-c,e-k (the Northeast Atlantic) (IBPWHM). ICES VOLUME 2 | ISSUE 95.

ICES. 2020. Working Group on Widely Distributed Stocks (WGWIDE 2020) Volume 2, Issue 82.

ICES (2021a). Workshop for the review of the assessment of a new rebuilding plan for western horse mackerel (WKWHMRP). ICES Scientific Reports. Copenhagen, ICES. Volume 3, Issue 37: 239 pp.

ICES (2021b). EU request to ICES on the assessment of a new rebuilding plan for western horse mackerel (Trachurus trachurus) in ICES Subarea 8 and divisions 2.a, 4.a, 5.b, 6.a, 7.a–c, and 7.e–k.In: ICES Special Request Advice, Ecoregions in the Northeast Atlantic sr.2021.04 (26 April 2021).

Pastoors, M. A., Campbell, A., Trijoulet, V., Skagen, D., Gras, M., Lambert, G., Sparrevohn, C. R., et al. 2020. Western Horse Mackerel Technical Focus Group on Harvest Control Rule Evaluations 2020. 43 pp pp.

Pastoors, M. A., A. Campbell, V. Trijoulet, M. Gras, G. Lambert, C. R. Sparrevohn, S. Mackinson and R. Ourens (2021). Updated diagnostics and simulations of the western horse mackerel rebuilding plan evaluations, ICES. Working document 01, WKWHMRP 2021: 97 pp.
